fix(lease): encrypt cached credentials in lease create command
Move find_encryption_provider, encrypt_credentials, and
decrypt_credentials to the shared lease module. Use encryption
in `fnox lease create` the same way `fnox exec` does, so cached
credentials are never stored as plaintext when an encryption
provider is configured.
