+++ src/lease.rs
@@ -1,5 +1,7 @@
+use crate::config::Config;
 use crate::env;
 use crate::error::{FnoxError, Result};
+use crate::providers::{self, ProviderCapability};
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
@@ -259,6 +261,70 @@ pub fn parse_duration(s: &str) -> Result<std::time::Duration> {
     }
 
     Ok(std::time::Duration::from_secs(total_secs))
+}
+
+/// Find an encryption provider if one is configured (default_provider with Encryption capability)
+pub async fn find_encryption_provider(
+    config: &Config,
+    profile: &str,
+) -> Option<(String, Box<dyn providers::Provider>)> {
+    let provider_name = match config.get_default_provider(profile) {
+        Ok(Some(name)) => name,
+        _ => return None,
+    };
+
+    let providers_map = config.get_providers(profile);
+    let provider_config = providers_map.get(&provider_name)?;
+
+    let provider =
+        match providers::get_provider_resolved(config, profile, &provider_name, provider_config)
+            .await
+        {
+            Ok(p) => p,
+            Err(e) => {
+                tracing::debug!(
+                    "Could not instantiate encryption provider '{}': {}",
+                    provider_name,
+                    e
+                );
+                return None;
+            }
+        };
+
+    if provider
+        .capabilities()
+        .contains(&ProviderCapability::Encryption)
+    {
+        Some((provider_name, provider))
+    } else {
+        None
+    }
+}
+
+/// Encrypt credential values using an encryption provider
+pub async fn encrypt_credentials(
+    provider: &dyn providers::Provider,
+    credentials: &HashMap<String, String>,
+) -> Result<HashMap<String, String>> {
+    let mut encrypted = HashMap::new();
+    for (key, value) in credentials {
+        let enc = provider.encrypt(value).await?;
+        encrypted.insert(key.clone(), enc);
+    }
+    Ok(encrypted)
+}
+
+/// Decrypt cached credential values using an encryption provider
+pub async fn decrypt_credentials(
+    provider: &dyn providers::Provider,
+    cached: &HashMap<String, String>,
+) -> Result<HashMap<String, String>> {
+    let mut decrypted = HashMap::new();
+    for (key, value) in cached {
+        let dec = provider.get_secret(value).await?;
+        decrypted.insert(key.clone(), dec);
+    }
+    Ok(decrypted)
 }
 
 #[cfg(test)]
